refactor: LaunchFolderAsync, LaunchFolderPathAsync, LaunchFileAsync, file-scoped namespaces
